properties order ios

+++ library/src/commonMain/kotlin/com/haitrvn/kam/AdError.kt
@@ -1,7 +1,7 @@
 package com.haitrvn.kam
 
 expect class AdError {
-    val code: Int
+    val code: Long
     val cause: AdError?
     val domain: String
     val message: String
